Routing problem and wavelength allocation
For a good performance of an optical network, it is necessary that its resources are treated properly, and this work seeks to address one of the main problems of optical networks: the routing and assignment of the wavelength (RWA). Many works found in literature address the problem of routing and wavelength assignment RWA (Routing and Wavelength Assigment) in optical networks WDM (Wavelength Division Multiplexing) in order to optimize the resources of the network resources, such as numbers of wavelengths and / or number of fibers in the network. Regarding connection requests, the problem can be of three types: static, adaptive and dynamic. In this present work, we have designed and implemented several algorithms to solve the static and adaptive RWA problem, where the main objective is to meet all network requests by the minimum path and assign the lowest number of wavelength, making the network more flexible and optimized. The computational results presented show the performance of each method.